requirements.txt
python-dotenv==0.13.0
python-youtube==0.6.0
tweepy==3.7.0
tqdm==4.46.1
pymongo==3.10.1